+++ docs/changes.rst
@@ -3,6 +3,8 @@ Weblate 4.4
 
 Not yet released.
 
+* Improved validation when creating component.
+
 Weblate 4.3.2
 -------------